requirements.txt
requests~=2.28.1
lxml~=4.9.1
Pillow~=9.2.0
pytesseract~=0.3.10
pytest~=7.2.0
responses~=0.22.0
